Audiolib: Change callbackval from uint32_t to intptr_t
Backported from NBlood.

From-SVN: r8427

+++ source/audiolib/include/fx_man.h
@@ -66,9 +66,9 @@ int FX_GetDevice(void);
 
 
 int FX_Play(char *ptr, uint32_t ptrlength, int loopstart, int loopend, int pitchoffset,
-                      int vol, int left, int right, int priority, fix16_t volume, uint32_t callbackval);
+                      int vol, int left, int right, int priority, fix16_t volume, intptr_t callbackval);
 int FX_Play3D(char *ptr, uint32_t ptrlength, int loophow, int pitchoffset, int angle,
-                  int distance, int priority, fix16_t volume, uint32_t callbackval);
+                  int distance, int priority, fix16_t volume, intptr_t callbackval);
 
 
 int FX_SetPrintf(void(*function)(const char *, ...));
@@ -87,7 +87,7 @@ static FORCE_INLINE int FX_CheckMVErr(int status)
     return status;
 }
 
-static FORCE_INLINE void FX_SetCallBack(void(*function)(uint32_t)) { MV_SetCallBack(function); }
+static FORCE_INLINE void FX_SetCallBack(void(*function)(intptr_t)) { MV_SetCallBack(function); }
 static FORCE_INLINE void FX_SetVolume(int volume) { MV_SetVolume(volume); }
 static FORCE_INLINE int FX_GetVolume(void) { return MV_GetVolume(); }
 static FORCE_INLINE void FX_SetReverb(int reverb) { MV_SetReverb(reverb); }
